修正 ヘッダーのアバター画像をグローバルで管理
